+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2022-05-01 :cake:
+
+- Publishes the first extension for Python Markdown of the repository:
+  `spantable`, custom Markdown table with support for `colspan` and `rowspan`
+- Improves the support for Material for MkDocs themes in `mkdocsoad`
+
 ## [0.0.1] - 2022-04-20 :sparkles:
 
 - First release with plugin for OpenAPI Documentation